feat(api): move event subscription logic to separate listener module (#314)
* feat(api): move event subscription logic to separate listener module

* chore: bump version to 0.2.15 and update changelog
