+++ blocklets/core/blocklet.yml
@@ -15,7 +15,7 @@ repository:
   type: git
   url: git+https://github.com/AIGNE-io/aigne-hub.git
 specVersion: 1.2.8
-version: 0.2.14
+version: 0.2.15
 logo: screenshots/logo.png
 files:
   - dist
