+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "ai-kit",
   "private": true,
-  "version": "0.2.14",
+  "version": "0.2.15",
   "workspaces": [
     "blocklets/*",
     "packages/*"
